fix asset_table bigdecimal to bigint

+++ src/main/java/com/idle/kb_i_dle_backend/domain/finance/dto/FinancialSumDTO.java
@@ -7,9 +7,9 @@ import lombok.Setter;
 @Getter
 @Setter
 public class FinancialSumDTO {
-    private BigDecimal amount;  // 금융 자산 합계
+    private Long amount;  // 금융 자산 합계
 
-    public FinancialSumDTO(BigDecimal amount) {
+    public FinancialSumDTO(Long amount) {
         this.amount = amount;
     }
 }
